documentando errores

+++ infrastructure/entry-points/reactive-web/src/main/java/com/peryloth/api/DTO/registry/RegistryUserDTO.java
@@ -1,7 +1,8 @@
 package com.peryloth.api.DTO.registry;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 
-public record RegistryUserDTO(Long id, String nombre, String apellido, String email, LocalDate fechaNacimiento,
+public record RegistryUserDTO(BigInteger id, String nombre, String apellido, String email, LocalDate fechaNacimiento,
                               String direccion, Long salarioBase) {
 }
